+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.1 - 2019-11-21
+### Fixed 
+- Fixed an issue where using `isOngoing()` wouldn't use the correct data and would crash. [#2](https://github.com/studioespresso/craft-date-range/issues/2)
+- Better date format parsing using ``DateTimeHelper::toDateTime()``
+
 ## 1.0.0 - 2019-10-24
 ### Added
 - Initial release 🎉